Restructure repository.
- Move tests folder to root path.
- Move lib out of MODIS to create shared space among all scripts.
- Simplify gdal_merge_test test helpers.
- Rename all folders to be lower case.
