+++ README.md
@@ -1,3 +1,4 @@
 # Snow Remote Sensing
+[![Build Status](https://travis-ci.org/UofU-Cryosphere/servir-snow.svg?branch=master)](https://travis-ci.org/UofU-Cryosphere/servir-snow)
 
 Helper scripts to download and process remote sensing data for MODIS and AMSR 2 satellites.
