+++ lib/__init__.py
@@ -1,5 +1,5 @@
-from .tile_file import TileFile
 from .source_folder import SourceFolder
+from .tile_file import TileFile
 from .tile_merger import TileMerger
 
 __all__ = [
