+++ modis/script_helpers.py
