+++ tests/modis/source_tiffs/bottom_right.tif
