+++ tests/modis/source_tiffs/middle.tif
