+++ tests/modis/source_tiffs/upper_left.tif
