+++ tests/modis/source_tiffs/upper_right.tif
